Ecomm-redirect (#1)
* feat: Organize JS files into src directory and update build

* feat: Implement e-commerce redirect logic

* refactor: Update ms-form-route.js and map for improved e-commerce handling and code organization

Refactored the ms-form-route.js to enhance the e-commerce redirect logic, consolidating event listeners and improving the structure for better maintainability. Updated the source map to reflect changes in the source files, ensuring accurate mapping for debugging and development purposes.

* feat: Add business bank account field mapping for HubSpot form submission

Enhanced the ms-form-route.js to include a new mapping for the "Do-you-have-a-business-bank-account" field, ensuring it is correctly submitted to HubSpot. Updated the source map and related files to reflect this addition, improving the overall functionality of the form submission process.